+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ---
 
+## [1.7.0] - 2025-08-XX
+
+- Aggregating namespace message "Initializing empty JSON"
+- Default logger name as "🌐Translate"
+
+
+---
+
 ## [1.6.0] - 2025-08-01
 
 - Added comprehensive logging system with debug/verbose levels